0) Full precision JSON output 1) Merged subprocess.py with cgifastrot.py as fastrot.py 2) Added replicate_results.py to recreate the contents of /data using the new code. These new results can be found within python-data/ 3) Added argparse
