+++ fastrot.py
@@ -23,7 +23,7 @@
    Modification History
    --------------------
    Mark Van Selous July 2021
-   - Rewrote cgifastrot.f in python as cgifastrot.py
+   - Rewrote cgifastrot.f and sublime.f in python as fastrot.py
    - Increased the latitude step size from 41 to 181.
    - Thanks to this increased step size and having more accurate trig functions
    than were available in fortran77, there is very slight alteration in the final results.
@@ -36,9 +36,7 @@
     inclination.
     - Initialized the parameters such that they can accept nine characters of data.
 """
-
 import math
-from subprocess import *
 import sys
 import json
 import logging
@@ -61,6 +59,137 @@ for idx in range(1, nb):
 speciesList = ['H2O', 'H2O_CH4', 'CO2', 'CO']
 
 logging.basicConfig(level=logging.DEBUG)
+
+# Constants
+sigma = 5.67e-5
+f0 = 1.39e6
+dynmm = 1.333e3
+boltz = 1.38e-16
+ergcal = 6.953e-17
+proton = 1.67e-24
+
+tstart = {'H2O': 190,
+          'H2O_CH4': 190,
+          'CO2': 100,
+          'CO': 60,
+          }
+
+
+def sublime(species, temp):
+    """
+    Description
+    -----------
+    Calculates the latent heat of sublimation and the vapor pressure of
+    the solid for various ices and the derivatives thereof
+
+    Parameters
+    ----------
+    species : str
+        Desired ice species to be considered. The valid inputs are:
+        - 'H2O'
+        - 'H20_CH4'
+        - 'CO2'
+        - 'CO'
+    temp: float
+        Temperature (in Kelvin) If temp <= 0 K, then the code defaults to a species
+        dependent initial value:
+        - H2O: 190 K
+        - H2O_CH4: 190 K
+        - CO2: 100 K
+        - CO: 60 K
+
+    Returns
+    -------
+    species : str
+    mass : float
+    xlt : float
+    xltprim : float
+    press : float
+    pprim : float
+    temp : float
+        The output temperature is either the input temperature
+        or the species dependent initial temperature if the input
+        temperature is less than 0 K.
+    """
+    mass, xlt, xltprim, press, pprim = None, None, None, None, None
+
+    # If temp <=0, then use the species dependent starting point.
+    if temp <= 0:
+        temp = tstart[species]
+
+    t = temp
+    t2 = t * t
+    t3 = t2 * t
+    t4 = t2 * t2
+    t5 = t4 * t
+    t6 = t3 * t3
+
+    if species == 'H2O':
+        # 100
+        mass = 18.
+        xlt = 12420. - 4.8 * t
+        xltprim = -4.8
+
+        # from Marti & Mauersberger (1993 GRL 20, 363)
+        p = -2663.5 / t + 12.537
+        p = 10. * 10. ** p
+        pp = (2663.5 / t2) * p
+
+        # mass, xlt, xltprim, press, pprim = p5100(mass, xlt, xltprim, p, pp)
+
+    elif species == 'H2O_CH4':
+        mass = 18.
+        xlt = 12160. + .5 * t - .033 * t2
+        xltprim = 0.5 - 0.066 * t
+
+        # from Marti & Mauersberger (1993 GRL 20, 363)
+        p = -2663.5 / t + 12.537
+        p = 10. * 10. ** p
+        pp = (+2663.5 / t2) * p
+
+        # mass, xlt, xltprim, press, pprim = p5100(mass, xlt, xltprim, p, pp)
+
+    elif species == 'CO2':
+        mass = 44.
+        xlt = 6269. + 9.877 * t - .130997 * t2 + 6.2735e-4 * t3 - 1.2699e-6 * t4
+        xltprim = 9.877 - .261994 * t + 1.88205e-3 * t2 - 5.0796e-6 * t3
+
+        p = 21.3807649e0 - 2570.647e0/t - 7.78129489e4/t2 + 4.32506256e6/t3 - 1.20671368e8/t4 + 1.34966306e9/t5
+        pp = 2570.647e0/t2 + 1.556258978e5/t3 - 12.97518768e6/t4 + 4.82685472e8/t5 - 6.7483153e9/t6
+        p = dynmm * 10. ** p
+        pp = pp * p
+
+        # mass, xlt, xltprim, press, pprim = p5100(mass, xlt, xltprim, p, pp)
+
+    elif species == 'CO':
+        mass = 28
+        if t > 68.127:
+            sys.exit(f'error in CO temp, T = {t}')
+        elif t > 61.544:
+            xlt = 1855 + 3.253 * t - .06833 * t2
+            xltprim = 3.253 - .13666 * t
+            p = 16.8655152e0 - 748.151471e0/t - 5.84330795e0/t2 + 3.93853859e0/t3
+            pp = 748.15147e0/t2 + 11.6866159e0/t3 - 11.81561577e0/t4
+        elif t >= 14.0:
+            xlt = 1893 + 7.331 * t + .01096 * t2 - .0060658 * t3 + 1.166e-4 * t4 - 7.8957e-7 * t5
+            xltprim = 7.331 + .02192 * t - .0181974 * t2 + 4.664e-4 * t3 - 3.94785e-6 * t4
+
+            p = 18.0741183e0 - 769.842078e0 / t - 12148.7759e0 / t2 + 2.7350095e5 / t3 - 2.9087467e6 / t4 + 1.20319418e7 / t5
+            pp = 769.842078e0 / t2 + 24297.5518 / t3 - 820502.85e0 / t4 + 11634986.8e0 / t5 - 60159709.e0 / t6
+            p = dynmm * 10. ** p
+            pp = pp * p
+        else:
+            sys.exit(f'error in CO temp, T= {t}')
+
+        # mass, xlt, xltprim, press, pprim = p5100(mass, xlt, xltprim, p, pp)
+
+    mass = mass * proton
+    xlt = xlt * ergcal
+    xltprim = xltprim * ergcal
+    press = p
+    pprim = pp
+
+    return species, mass, xlt, xltprim, press, pprim, temp
 
 
 def run_model(species, a_v, a_ir, rh, inc, temp=-1):
@@ -178,11 +307,11 @@ def run_model(species, a_v, a_ir, rh, inc, temp=-1):
         "Species": species,
         "Obliquity": inc,
         "r_H": rh,
-        "rlog": "{:.3f}".format(rlog),
+        "rlog": rlog,
         "a_v": a_v,
         "a_ir": a_ir,
-        "Zbar": "{:.3e}".format(zbar),
-        "Zlog": "{:.3f}".format(zlog),
+        "Zbar": zbar,
+        "Zlog": zlog,
     }))
 
     return species, inc, rh, rlog, a_v, a_ir, zbar, zlog
@@ -196,7 +325,8 @@ def main_loop(n, species, a_v, a_ir, rh, incl, temp, root, nflag, perc, gd):
         counter (ranging from 0 to nb-1) which tracks
     species: str
         Inputted species
-    a_v : float (a_v > 0)
+    a_v : float
+        (a_v > 0)
         Inputted visual albedo
     a_ir : float
         Inputted infrared albedo
@@ -227,7 +357,7 @@ def main_loop(n, species, a_v, a_ir, rh, incl, temp, root, nflag, perc, gd):
     temp : float
         Temperature (in Kelvin). This is one of the three parameters which is updated by main_loop().
     gd : None / float
-        gd is used in the case of nflag >= 10,000,000
+        gd is used in the case of nflag >= 5,000
     perc : float
     """
     root_t = math.sqrt(temp)
@@ -274,7 +404,7 @@ def main_loop(n, species, a_v, a_ir, rh, incl, temp, root, nflag, perc, gd):
     if abs(phi / sun) < 1e-4 or abs(phi) < 1e-4:
         gd, perc = calc_perc(n, rh, incl, perc)
         return temp, gd, perc, nflag
-    if nflag >= 1000:
+    if nflag >= 5000:
         subdis = z[gd] * 4 * math.pi * rh * 149.6e11
         if perc != 0:
             if (subdis / perc) < 1e-02:
@@ -300,12 +430,26 @@ def calc_perc(n, rh, incl, perc):
 
 
 ############
-a_v_in = 0.05
-a_ir_in = 0.00
-rh_in = 3.98
-inc_in = 90
 
-# speciesList = ['H2O', 'H2O_CH4', 'CO2', 'CO']
-index_in = speciesList.index('H2O') + 1
 
-run_model(index_in, a_v_in, a_ir_in, rh_in, inc_in, )
+if __name__ == '__main__':
+    import argparse
+
+    parser = argparse.ArgumentParser(description="NEED description", formatter_class=argparse.RawTextHelpFormatter,)
+    parser.add_argument('species', type=str,
+                        help="Desired ice species to be considered. "
+                             "The valid inputs are: 'H2O', 'H20_CH4', 'CO2', and 'CO'")
+    parser.add_argument('--a_v', metavar='visual albedo', type=float, required=True,)
+    parser.add_argument('--a_ir', metavar='infrared albedo', type=float, required=True,)
+    parser.add_argument('--rh', metavar='heliocentric_distance', type=float, required=True,)
+    parser.add_argument('--inc', metavar='obliquity', type=float, required=True,)
+    parser.add_argument('--temp', metavar='temperature', type=float, default=-1,
+                        help="Not passing a starting temperature will default to a species dependent starting value:\n"
+                             "H2O: 190 K\n"
+                             "H2O_CH4: 190 K\n"
+                             "CO2: 100 K\n"
+                             "CO: 60 K\n")
+
+    args = parser.parse_args()
+    print(args)
+    results = run_model(args.species, args.a_v, args.a_ir, args.rh, args.inc, args.temp)
